Change gen.py Acceptor to use match, not search
This makes it easier to do the common thing:

    -W malloc  // only matches malloc

But also allows the uncommon thing:

    -W ".*foo.*"  // matches any symbol with "foo" in it

I also modified the rot13 integration test to use this instead of
wrapping malloc, free and memcpy.

Finally, I changed gen.py to default to including all symbols in the
current file, regardless of the whitelist/blacklist. This is commonly
what you want if you wrote that header by hand. If not, you can always
explicitly blacklist the header and whitelist the symbols you want.