Remove Type.equals and move it to test code
As far as I can tell, testing types for equality will not be useful in
client code. Moving it out of naclbind.js reduces the API footprint.

The current implementation is also incorrect; it doesn't work properly
with recursive types. The new one in test/js/equals.js properly tracks
types so it doesn't infinitely recurse.