Use underlying_typedef_type instead of get_canonical
I didn't see this function before because it is defined on Cursor
instead of Type. It is necessary to use this, however to define all
typedefs in the correct order.

The test I added exercises this issue that occurs with FILE, basically:

    typedef struct __sFILE __FILE;
    struct __sFILE {
      __FILE* f;
    };
    typedef __FILE FILE;

get_canonical causes us to map FILE -> __sFILE and skip __FILE.