Use a hashmap for handles instead of a sorted list
This uses a technique from the lua table implementation called Brent's
variation. Every key has a "main position", which is the index that the
hash function generates. If there is a hash collision, check the entry
at that spot; if it is not in its main position, move it to the next
free entry and use the main position for the newly added entry.