Long types now used "tagged arrays"
I want to be able to pass types to the NaCl module that are not natively
supported by Pepper (such as int64). Previously I used an array of two
elements for this, but now that I need more than just a 64-bit value
(e.g. a function value), I need to pass a tag as well.

So a 64-bit int is now ['long', <low bits>, <hi bits>].