Add FunctionUntyped, compatible with all fn types
We want to be able to pass JavaScript functions as callbacks to C
functions in the NaCl module. They must be stored in a handle, and
therefore must be typed. FunctionUntyped is a new type that is
compatible with the Function and FunctionNoProto types. This means that
any JavaScript function can be passed as a function pointer to any C
function.