Rename message -> request in C files
This name is more accurate, in that there can be multiple types of
messages: Requests and Responses.

This change is in preparation for allowing the C code to create and
parse Responses. This is necessary to send Responses (in response to an
initial Request from JavaScript) as well as receive Responses (in
response to a callback returning from JavaScript).